#157030

Related work items: #157030

+++ Logitude/Simplog.Data/CommonDataModel/EntityPOCOs/HTSCode.cs
@@ -19,8 +19,8 @@ namespace Simplog.Data.CommonDataModel.EntityPOCOs
         public bool ApprovedByCustomer { get; set; }
         public bool InActive { get; set; }
         public int? LineNumber { get; set; }
-        public decimal? VATPercentage { get; set; }
-        public decimal? DutiesPercentage { get; set; }
+        public double? VATPercentage { get; set; }
+        public double? DutiesPercentage { get; set; }
         public string OtherDuties { get; set; }
         public string Remarks { get; set; }
     }
